documentation: написал JavaDoc-комментарии для классов работающих с Spring Security
